voyager-backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
mysql-connector-python==8.4.0
bcrypt==4.1.3
PyJWT==2.8.0
